+++ src/main/java/com/pyramid/model/Cart.java
@@ -3,14 +3,14 @@ package com.pyramid.model;
 public class Cart {
 	
 	private String p_name;
-	private double p_qty;
+	private int p_qty;
 	private double total_p_price;
 	
 	
 	public Cart() {
 		
 	}
-	public Cart(String p_name, double p_qty, double total_p_price) {
+	public Cart(String p_name, int p_qty, double total_p_price) {
 		super();
 		this.p_name = p_name;
 		this.p_qty = p_qty;
@@ -22,10 +22,10 @@ public class Cart {
 	public void setP_name(String p_name) {
 		this.p_name = p_name;
 	}
-	public double getP_qty() {
+	public int getP_qty() {
 		return p_qty;
 	}
-	public void setP_qty(double p_qty) {
+	public void setP_qty(int p_qty) {
 		this.p_qty = p_qty;
 	}
 	public double getTotal_p_price() {
